backend/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
razorpay==1.3.0
gunicorn==21.2.0
email-validator==2.0.0
Werkzeug==2.3.7 